Passo 4 Implementação do Sistema de Level - inventado por mim
Implementei o sistema de Level, se baseando nos sistemas criados.

Onde para upar o Level você gasta ponto, e quando chega em um certo nível seu ponto por click multiplica por algum valor, futuramente planejo implementar um sistema de bosses onde para liberar o boss terá que ter certo nível.

Também fiz algumas alterações em nome de funções e coisas pequenas.

+++ Criando jogos/Jogo número 1 - Clicker lizard/JS/clickerButtonPontsGroup.js
@@ -5,7 +5,7 @@ export function clickButtonPonts() {
     // Campo dos Pontos
     let spanPontos = document.getElementById("spanPontos")
     // Pontos
-    let pontos = parseInt(spanPontos.textContent)
+    let pontos = parseFloat(spanPontos.textContent)
     pontos += getSumPont()
 
     spanPontos.textContent = pontos
@@ -28,8 +28,8 @@ export function getPonts() {
     return document.getElementById("spanPontos");
 }
 
-// Pegar valor inteiro de Pontos atuais
+// Pegar valor inteiro de Pontos atuais (depois alterado para Float)
 export function getIntPonts() {
     let valuePonts = document.getElementById("spanPontos").textContent
-    return parseInt(valuePonts);
+    return parseFloat(valuePonts);
 }
